refactor(pytorch): extract checkpoint lifecycle
